test: move scanner regression into its own file
The pre-existing GitHub-hitting tests in bun-create.test.ts fail under
restricted network (TLS cert errors in the sandbox), which left the
whole-file pass/fail signal red regardless of the fix. Pull the new
regression out so it can be evaluated on its own.
